+++ tests/integration/config_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/sorintlab/stolon/internal/common"
 	"github.com/sorintlab/stolon/internal/store"
 
-	"github.com/gofrs/uuid"
+	"github.com/google/uuid"
 )
 
 func TestServerParameters(t *testing.T) {
@@ -53,7 +53,7 @@ func TestServerParameters(t *testing.T) {
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 	defer tstore.Stop()
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	storePath := filepath.Join(common.StorePrefix, clusterName)
 
@@ -150,7 +150,7 @@ func TestWalLevel(t *testing.T) {
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 	defer tstore.Stop()
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	storePath := filepath.Join(common.StorePrefix, clusterName)
 
@@ -265,7 +265,7 @@ func TestWalKeepSegments(t *testing.T) {
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 	defer tstore.Stop()
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	storePath := filepath.Join(common.StorePrefix, clusterName)
 
@@ -438,7 +438,7 @@ func TestAlterSystem(t *testing.T) {
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 	defer tstore.Stop()
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	storePath := filepath.Join(common.StorePrefix, clusterName)
 
@@ -510,7 +510,7 @@ func TestAdditionalReplicationSlots(t *testing.T) {
 	}
 	defer os.RemoveAll(dir)
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	tks, tss, tp, tstore := setupServers(t, clusterName, dir, 2, 1, false, false, nil)
 	defer shutdown(tks, tss, tp, tstore)
@@ -659,7 +659,7 @@ func TestAutomaticPgRestart(t *testing.T) {
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 	defer tstore.Stop()
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	storePath := filepath.Join(common.StorePrefix, clusterName)
 
@@ -783,7 +783,7 @@ func TestAdvertise(t *testing.T) {
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 	defer tstore.Stop()
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	storePath := filepath.Join(common.StorePrefix, clusterName)