+++ tests/integration/init_test.go
@@ -24,7 +24,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/gofrs/uuid"
+	"github.com/google/uuid"
 	"github.com/sorintlab/stolon/internal/cluster"
 	"github.com/sorintlab/stolon/internal/common"
 	"github.com/sorintlab/stolon/internal/store"
@@ -44,7 +44,7 @@ func TestInit(t *testing.T) {
 
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	initialClusterSpec := &cluster.ClusterSpec{
 		InitMode:           cluster.ClusterInitModeP(cluster.ClusterInitModeNew),
@@ -93,7 +93,7 @@ func TestInitNewNoMerge(t *testing.T) {
 }
 
 func testInitNew(t *testing.T, merge bool) {
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	dir, err := os.MkdirTemp("", "")
 	if err != nil {
@@ -172,7 +172,7 @@ func TestInitExistingNoMerge(t *testing.T) {
 }
 
 func testInitExisting(t *testing.T, merge bool) {
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	dir, err := os.MkdirTemp("", "")
 	if err != nil {
@@ -318,7 +318,7 @@ func TestInitUsers(t *testing.T) {
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 
 	// Test pg-repl-username == pg-su-username but password different
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 	tk, err := NewTestKeeper(t, dir, clusterName, "user01", "password01", "user01", "password02", tstore.storeBackend, storeEndpoints)
 	if err != nil {
 		t.Fatalf("unexpected err: %v", err)
@@ -332,7 +332,7 @@ func TestInitUsers(t *testing.T) {
 	}
 
 	// Test pg-repl-username == pg-su-username
-	clusterName = uuid.Must(uuid.NewV4()).String()
+	clusterName = uuid.NewString()
 	storePath := filepath.Join(common.StorePrefix, clusterName)
 
 	sm := store.NewKVBackedStore(tstore.store, storePath)
@@ -375,7 +375,7 @@ func TestInitUsers(t *testing.T) {
 	}
 
 	// Test pg-repl-username != pg-su-username and pg-su-password defined
-	clusterName = uuid.Must(uuid.NewV4()).String()
+	clusterName = uuid.NewString()
 	storePath = filepath.Join(common.StorePrefix, clusterName)
 
 	sm = store.NewKVBackedStore(tstore.store, storePath)
@@ -421,7 +421,7 @@ func TestInitialClusterSpec(t *testing.T) {
 	tstore := setupStore(t, dir)
 	defer tstore.Stop()
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 	storePath := filepath.Join(common.StorePrefix, clusterName)
@@ -485,9 +485,9 @@ func TestExclusiveLock(t *testing.T) {
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 	defer tstore.Stop()
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
-	u := uuid.Must(uuid.NewV4())
+	u := uuid.New()
 	id := fmt.Sprintf("%x", u[:4])
 
 	tk1, err := NewTestKeeperWithID(t, dir, id, clusterName, pgSUUsername, pgSUPassword, pgReplUsername, pgReplPassword, tstore.storeBackend, storeEndpoints)
@@ -541,9 +541,9 @@ func TestPasswordTrailingNewLine(t *testing.T) {
 	storeEndpoints := fmt.Sprintf("%s:%s", tstore.listenAddress, tstore.port)
 	defer tstore.Stop()
 
-	clusterName := uuid.Must(uuid.NewV4()).String()
+	clusterName := uuid.NewString()
 
-	u := uuid.Must(uuid.NewV4())
+	u := uuid.New()
 	id := fmt.Sprintf("%x", u[:4])
 
 	pgSUPassword := "stolon_superuserpassword\n"