+++ tests/integration/utils.go
@@ -42,7 +42,7 @@ import (
 	pg "github.com/sorintlab/stolon/internal/postgresql"
 	"github.com/sorintlab/stolon/internal/store"
 
-	"github.com/gofrs/uuid"
+	"github.com/google/uuid"
 	_ "github.com/lib/pq"
 )
 
@@ -589,7 +589,7 @@ func NewTestKeeperWithID(t *testing.T, dir, uid, clusterName, pgSUUsername, pgSU
 }
 
 func NewTestKeeper(t *testing.T, dir, clusterName, pgSUUsername, pgSUPassword, pgReplUsername, pgReplPassword string, storeBackend store.Backend, storeEndpoints string, a ...string) (*TestKeeper, error) {
-	u := uuid.Must(uuid.NewV4())
+	u := uuid.New()
 	uid := fmt.Sprintf("%x", u[:4])
 
 	return NewTestKeeperWithID(t, dir, uid, clusterName, pgSUUsername, pgSUPassword, pgReplUsername, pgReplPassword, storeBackend, storeEndpoints, a...)
@@ -894,7 +894,7 @@ type TestSentinel struct {
 }
 
 func NewTestSentinel(t *testing.T, dir string, clusterName string, storeBackend store.Backend, storeEndpoints string, a ...string) (*TestSentinel, error) {
-	u := uuid.Must(uuid.NewV4())
+	u := uuid.New()
 	uid := fmt.Sprintf("%x", u[:4])
 
 	args := []string{}
@@ -933,7 +933,7 @@ type TestProxy struct {
 }
 
 func NewTestProxy(t *testing.T, dir string, clusterName, pgSUUsername, pgSUPassword, pgReplUsername, pgReplPassword string, storeBackend store.Backend, storeEndpoints string, a ...string) (*TestProxy, error) {
-	u := uuid.Must(uuid.NewV4())
+	u := uuid.New()
 	uid := fmt.Sprintf("%x", u[:4])
 
 	listenAddress, port, err := getFreePort(true, false)
@@ -1117,7 +1117,7 @@ func NewTestStore(t *testing.T, dir string, a ...string) (*TestStore, error) {
 }
 
 func NewTestEtcd(t *testing.T, dir string, backend store.Backend, a ...string) (*TestStore, error) {
-	u := uuid.Must(uuid.NewV4())
+	u := uuid.New()
 	uid := fmt.Sprintf("%x", u[:4])
 
 	dataDir := filepath.Join(dir, fmt.Sprintf("etcd%s", uid))